+++ tests/live_apps/run.py
@@ -4,14 +4,8 @@ from importlib import import_module
 
 from werkzeug.serving import generate_adhoc_ssl_context
 from werkzeug.serving import run_simple
-from werkzeug.serving import WSGIRequestHandler
 from werkzeug.wrappers import Request
 from werkzeug.wrappers import Response
-
-
-class WSGI11RequestHandler(WSGIRequestHandler):
-    protocol_version = "HTTP/1.1"
-
 
 name = sys.argv[1]
 mod = import_module(f"{name}_app")
@@ -29,9 +23,6 @@ kwargs = getattr(mod, "kwargs", {})
 kwargs.update(hostname="127.0.0.1", port=5000, application=app)
 kwargs.update(json.loads(sys.argv[2]))
 ssl_context = kwargs.get("ssl_context")
-
-if kwargs.get("request_handler") == "HTTP/1.1":
-    kwargs["request_handler"] = WSGI11RequestHandler
 
 if ssl_context == "custom":
     kwargs["ssl_context"] = generate_adhoc_ssl_context()